Exception handling - delete

+++ src/main/java/com/carlosborges/services/UserService.java
@@ -4,10 +4,12 @@ import java.util.List;
 import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.stereotype.Service;
 
 import com.carlosborges.entities.User;
 import com.carlosborges.repositories.UserRepository;
+import com.carlosborges.services.exceptions.DatabaseException;
 import com.carlosborges.services.exceptions.ResourceNotFoundException;
 
 @Service
@@ -30,7 +32,16 @@ public class UserService {
 	}
 	 
 	public void delete(Long id) {
-		repository.deleteById(id);
+		Boolean del = repository.existsById(id);
+		if(del == false) {
+			 throw new ResourceNotFoundException(id);
+		} 
+		try {
+			repository.deleteById(id);	
+		}catch (DataIntegrityViolationException e) {
+			throw new DatabaseException(e.getMessage());
+		}
+		
 	}
 	public User update(Long id, User obj) {
 		User entity = repository.getReferenceById(id);
